requirements.txt
celery>=5.4.0,<6
Django>=5.2,<6
django-celery-beat>=2.9.0
django-celery-results>=2.5.1
django-cors-headers>=4.4.0
django-crispy-forms>=2.4
django-environ>=0.12.0
django-import-export>=4.0.0
django-prometheus>=2.3.1
django-redis>=5.4.0
django-unfold>=0.68.0
djangorestframework>=3.16.1
psycopg2-binary>=2.9.10
redis>=7.0.0
requests>=2.32.0
polars>=1.36.0
tqdm>=4.66.0

# Machine Learning
scikit-learn>=1.5.0
xgboost>=2.1.0
lightgbm>=4.5.0
numpy>=2.0.0
pandas>=2.2.0
joblib>=1.4.0
